requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==1.10.8
googlesearch-python==1.2.3 